Refactor: Split DemurrageCalculatorDialog into smaller files
Split the DemurrageCalculatorDialog component into smaller, more manageable files to improve code organization and readability.
